+++ task/lib/src/widgets/add_task_dialog.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:uuid/uuid.dart';
 
 import 'package:taskc/json.dart';
+import 'package:taskw/taskw.dart';
 
 import 'package:task/task.dart';
 
@@ -28,17 +29,26 @@ class AddTaskDialog extends StatelessWidget {
         ),
         ElevatedButton(
           onPressed: () async {
-            var now = DateTime.now().toUtc();
-            StorageWidget.of(context).mergeTask(
-              Task(
-                status: 'pending',
-                uuid: Uuid().v1(),
-                entry: now,
-                description: controller.text,
-                modified: now,
-              ),
-            );
-            Navigator.of(context).pop();
+            try {
+              validateTaskDescription(controller.text);
+              var now = DateTime.now().toUtc();
+              StorageWidget.of(context).mergeTask(
+                Task(
+                  status: 'pending',
+                  uuid: Uuid().v1(),
+                  entry: now,
+                  description: controller.text,
+                  modified: now,
+                ),
+              );
+              Navigator.of(context).pop();
+            } on FormatException catch (e, trace) {
+              showExceptionDialog(
+                context: context,
+                e: e,
+                trace: trace,
+              );
+            }
           },
           child: Text('Submit'),
         ),