+++ taskw/lib/taskw.dart
@@ -9,3 +9,4 @@ export 'src/modify.dart';
 export 'src/profiles.dart';
 export 'src/task_copy_with.dart';
 export 'src/urgency.dart';
+export 'src/validate.dart';