+++ taskw/test/validate_test.dart
@@ -1,0 +1,20 @@
+import 'package:test/test.dart';
+
+import 'package:taskw/taskw.dart';
+
+void main() {
+  test('test validation of task description', () {
+    expect(
+      () => validateTaskDescription('foo'),
+      returnsNormally,
+    );
+    expect(
+      () => validateTaskDescription(r'foo\'),
+      throwsException,
+    );
+    expect(
+      () => validateTaskDescription(r'\'),
+      throwsException,
+    );
+  });
+}